pyproject.toml
[tool.poetry]
name = "mspy-account"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.13.1"
django = "^4.0"
django-cors-headers = "^3.11.0"
djangorestframework-camel-case = "^1.3.0"
python-dateutil = "^2.8.2"
uvicorn = "^0.21.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
whitenoise = "^6.5.0"
pydantic = "^2.5.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
